+++ omisego-sdk/src/main/java/co/omisego/omisego/websocket/OMGSocketClient.kt
@@ -26,14 +26,14 @@ import co.omisego.omisego.websocket.channel.dispatcher.delegator.SocketDelegator
 import co.omisego.omisego.websocket.channel.dispatcher.delegator.SocketReceiveParser
 import co.omisego.omisego.websocket.enum.SocketStatusCode
 import co.omisego.omisego.websocket.interval.SocketHeartbeat
-import co.omisego.omisego.websocket.listener.CompositeSocketChannelListener
-import co.omisego.omisego.websocket.listener.CompositeSocketConnectionListener
 import co.omisego.omisego.websocket.listener.SocketChannelListener
-import co.omisego.omisego.websocket.listener.SocketChannelListenerSet
 import co.omisego.omisego.websocket.listener.SocketConnectionListener
-import co.omisego.omisego.websocket.listener.SocketConnectionListenerSet
 import co.omisego.omisego.websocket.listener.SocketCustomEventListener
-import co.omisego.omisego.websocket.listener.SocketCustomEventListenerSet
+import co.omisego.omisego.websocket.listener.internal.CompositeSocketChannelListener
+import co.omisego.omisego.websocket.listener.internal.CompositeSocketConnectionListener
+import co.omisego.omisego.websocket.listener.internal.SocketChannelListenerSet
+import co.omisego.omisego.websocket.listener.internal.SocketConnectionListenerSet
+import co.omisego.omisego.websocket.listener.internal.SocketCustomEventListenerSet
 import okhttp3.OkHttpClient
 import okhttp3.Request
 import okhttp3.WebSocket
