+++ omisego-sdk/src/main/java/co/omisego/omisego/websocket/channel/SocketChannel.kt
@@ -14,13 +14,13 @@ import co.omisego.omisego.websocket.channel.SocketChannelContract.Dispatcher
 import co.omisego.omisego.websocket.channel.SocketChannelContract.SocketClient
 import co.omisego.omisego.websocket.enum.SocketEventSend
 import co.omisego.omisego.websocket.enum.SocketStatusCode
-import co.omisego.omisego.websocket.listener.CompositeSocketChannelListener
-import co.omisego.omisego.websocket.listener.CompositeSocketConnectionListener
+import co.omisego.omisego.websocket.listener.internal.CompositeSocketChannelListener
+import co.omisego.omisego.websocket.listener.internal.CompositeSocketConnectionListener
 import co.omisego.omisego.websocket.listener.SocketChannelListener
-import co.omisego.omisego.websocket.listener.SocketChannelListenerSet
+import co.omisego.omisego.websocket.listener.internal.SocketChannelListenerSet
 import co.omisego.omisego.websocket.listener.SocketConnectionListener
-import co.omisego.omisego.websocket.listener.SocketConnectionListenerSet
-import co.omisego.omisego.websocket.listener.SocketCustomEventListenerSet
+import co.omisego.omisego.websocket.listener.internal.SocketConnectionListenerSet
+import co.omisego.omisego.websocket.listener.internal.SocketCustomEventListenerSet
 import java.util.concurrent.ArrayBlockingQueue
 import java.util.concurrent.BlockingQueue
 import java.util.concurrent.TimeUnit
