+++ omisego-sdk/src/main/java/co/omisego/omisego/websocket/channel/SocketChannelContract.kt
@@ -8,11 +8,11 @@ package co.omisego.omisego.websocket.channel
  */
 
 import co.omisego.omisego.model.socket.SocketSend
-import co.omisego.omisego.websocket.listener.CompositeSocketChannelListener
-import co.omisego.omisego.websocket.listener.CompositeSocketConnectionListener
+import co.omisego.omisego.websocket.listener.internal.CompositeSocketChannelListener
+import co.omisego.omisego.websocket.listener.internal.CompositeSocketConnectionListener
 import co.omisego.omisego.websocket.SocketClientContract
 import co.omisego.omisego.websocket.enum.SocketStatusCode
-import co.omisego.omisego.websocket.listener.SocketCustomEventListenerSet
+import co.omisego.omisego.websocket.listener.internal.SocketCustomEventListenerSet
 import java.util.concurrent.BlockingQueue
 import java.util.concurrent.atomic.AtomicBoolean
