+++ omisego-sdk/src/main/java/co/omisego/omisego/websocket/listener/internal/SocketConnectionListenerSet.kt
@@ -1,4 +1,6 @@
-package co.omisego.omisego.websocket.listener
+package co.omisego.omisego.websocket.listener.internal
+
+import co.omisego.omisego.websocket.listener.SocketConnectionListener
 
 /**
  * OmiseGO
